+++ CHANGELOG.md
@@ -5,8 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
+* fix: when moving pane focus off screen edge to the next tab, the pane on the screen edge is now focused (https://github.com/zellij-org/zellij/pull/2293)
+* fix: adding panes to lone stack (https://github.com/zellij-org/zellij/pull/2298)
+
+## [0.35.2] - 2023-03-10
 * fix: get "zellij attach --create" working again (https://github.com/zellij-org/zellij/pull/2247)
 * fix: crash when closing tab with command panes (https://github.com/zellij-org/zellij/pull/2251)
+* Terminal compatibility: pad end of line on `CSI P` (https://github.com/zellij-org/zellij/pull/2259)
 
 ## [0.35.1] - 2023-03-07
 * fix: show visual error when unable to split panes vertically/horizontally (https://github.com/zellij-org/zellij/pull/2025)